jwt required imported it

+++ resources/item.py
@@ -1,7 +1,6 @@
-import uuid
-from flask import request
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort
+from flask_jwt_extended import jwt_required, get_jwt
 from sqlalchemy.exc import SQLAlchemyError
 
 from db import db
@@ -13,12 +12,19 @@ blp = Blueprint("Items", __name__, description="Operations on items")
 
 @blp.route("/item/<int:item_id>")
 class Item(MethodView):
+    @jwt_required()
     @blp.response(200, ItemSchema)
     def get(self, item_id):
         item = ItemModel.query.get_or_404(item_id) #get or 404 é auto explicativo (flask-sqlalchemy)
         return item
 
+    @jwt_required()
     def delete(self, item_id):
+        # ----------optional jwt claims verification --------
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin privilege required.")
+        # ----------optional jwt claims verification --------
         item = ItemModel.query.get_or_404(item_id)
         db.session.delete(item)
         db.session.commit()
@@ -44,12 +50,14 @@ class Item(MethodView):
 
 @blp.route("/item")
 class ItemList(MethodView):
+    @jwt_required()
     @blp.response(200, ItemSchema(many=True)) # many = true, cria uma lista de itens.
     def get(self):
         return ItemModel.query.all()
     
     # ItemSchema não está apenas validando o tipo de dado recebido, mas tambem fazendo:
     # item_data = request.get_json(), ou seja um "auto-fetch"
+    @jwt_required() # post requer um token jwt
     @blp.arguments(ItemSchema)
     @blp.response(201, ItemSchema)
     def post(self, item_data):
